remake all processes to services
