+++ DataReader.Application/ModelsServices/LogsService.cs
@@ -1,0 +1,7 @@
+﻿namespace DataReader.Application.Services
+{
+  public class LogsService
+  {
+
+  }
+}
