+++ DataReader.Application/ModelsServices/ProjectsService.cs
@@ -1,0 +1,7 @@
+﻿namespace DataReader.Application.Services
+{
+  public class ProjectsService
+  {
+
+  }
+}
