+++ DataReader.Application/ModelsServices/WorkItemsService.cs
@@ -1,0 +1,6 @@
+﻿namespace DataReader.Application.Services
+{
+  public class WorkItemsService
+  {
+  }
+}
