--- DataReader.Core/Abstractions/UseCases/IRead.cs
@@ -1,7 +1,0 @@
-﻿namespace DataReader.Core.Abstractions.UseCases
-{
-    public interface IRead
-    {
-
-    }
-}
