--- DataReader.Core/Abstractions/UseCases/ISync.cs
@@ -1,9 +1,0 @@
-﻿using DataReader.Core.Contracts;
-
-namespace DataReader.Core.Abstractions.UseCases
-{
-  public interface ISync
-  {
-    Task Synchronisation(ContractsWrapper requests);
-  }
-}
