--- DataReader.Core/Abstractions/UseCases/IWrite.cs
@@ -1,7 +1,0 @@
-﻿namespace DataReader.Core.Abstractions.UseCases
-{
-    public interface IWrite
-    {
-
-    }
-}
